level-paramset creation after requirement creation overhaul
this was damn hard to get right. for the relations to become properly
serialized it was necessary to

1. create and save the level-paramsets
2. add one level-paramset to its level
3. save the level-paramset
4. save the level
5. repeat to 2-4 for all levels
6. add one level-paramsets to requirement
7. save level-paramsets
8. save the requirement

in this order. especially with a OneToMany relation it is always
necessary to add with pushObject to the many part. save the model on the
one side and only afterwards save the model on the many side.