Made array-like data-flow more robust
calling notifyPropertyChange on a reduceComputed property forces a total
recalculation, so this had to be avoided. At the moment, we use another
boolean property updateTrigger, which we switch once an update occurs.

the reduceComputed calls became more robust in that they check, whether
all data is already available.

a few console.logs show the problem of level's Em.computed.map to miss
out on necessary updates.